Issue: WSWRPR-1314 EAR 3.0 release for RT + Firecrest
Resolving build IAR build warning reported by Dapeng on the latest build.

Reviewers: Abhinav K, Sandip D, Supriya I.

+++ wifidriver/mlan_cmdevt.c
@@ -880,7 +880,7 @@ done:
        implementation here.
     */
 
-    return wifi_prepare_and_send_cmd(pmpriv, cmd_no, cmd_action, cmd_oid, pioctl_buf, pdata_buf, (int)pmpriv->bss_type,
+    return wifi_prepare_and_send_cmd(pmpriv, cmd_no, cmd_action, cmd_oid, pioctl_buf, pdata_buf, pmpriv->bss_type,
                                      NULL);
 }
